% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKm02q0.R
\name{AKm02q0}
\alias{AKm02q0}
\title{AKm02q0 derive m0 from q0}
\usage{
AKm02q0(m0, constant, slope)
}
\arguments{
\item{m0}{a value of m0, per the period lifetable derivation}

\item{constant}{the intercept of the relevant Andreev-Kingkade segment}

\item{slope}{the slope of the relevant Andreev-Kingkade segment}
}
\value{
q0 the estimate of q0 according to the identity between a0, m0, q0
}
\description{
Derive m0 from q0 according to the relevant segment of the Andreeve-Kingkade formula. This is elegant because it's an analytic solution, but ugly because, man, look at it. Carl Boe got this from Maple I think. This formula is only necessary because AK start with q0 whereas the HMD starts with m0, so we needed to adapt.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}
