% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDa0.R
\name{CDa0}
\alias{CDa0}
\title{\code{CDa0} estimate a0 using the Coale-Demeny rule of thumb.}
\usage{
CDa0(m0, sex = "f")
}
\arguments{
\item{m0}{a value or vector of values of m0, the death rate for age 0 infants.}

\item{sex}{either "m" or "f"}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of a_0 values.
}
\description{
\code{CDa0} is an auxiliary function used by the four HMD lifetable functions, \code{ltper_AxN()}, \code{ltcoh_AxN()}, \code{ltperBoth_AxN()}, \code{ltcohBoth_AxN()}, described in formulas 61 and 62 of MP version 5.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

