% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRuncoh.R
\name{checkRuncoh}
\alias{checkRuncoh}
\title{\code{checkRuncoh} a function to do an automatic check for whether cohort lifetables are run}
\usage{
checkRuncoh(WORKING = getwd(), males = NULL, females = NULL,
  OPENAGE = 110, save.bin = TRUE, XXX = NULL, LDBPATH = NULL,
  IDBPATH = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD country abbreviation. Default \code{getwd()}.}

\item{males}{the cohort components \code{data.frame} object, as created by \code{getCohortComponents()} for males. If left as \code{NULL}, \code{getCohortComponents()} is called.}

\item{females}{the cohort components \code{data.frame} object, as created by getCohortComponents() for females. If left as \code{NULL}, \code{getCohortComponents()} is called.}

\item{OPENAGE}{the desired open age. Default value is 110}

\item{save.bin}{logical. Default \code{TRUE}. Should the output be saved as \code{Rbin/checkRuncoh.Rdata} as well?}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is extracted from \code{WORKING} as the past path part.}

\item{LDBPATH}{in case the LexisDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "LexisDB")}}

\item{IDBPATH}{in case the InputDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "InputDB")}}
}
\value{
a list with
\itemize{
 \item{"timestamp"}{e.g. "18 Dec 2012" *character class), just so you know whether the check was recent or not.}
 \item{"m"}{logical, males appear to have at least 1 extinct cohort}
 \item{"f"}{logical, females appear to have at least 1 extinct cohort}
 \item{"both"}{logical, both males and females appear to have at least 1 extinct cohort}
 \item{"either"}{logical, either males or females appear to have at least 1 extinct cohort}
 \item{"note"}{"TRUE indicates that a cohort appears extinct"}
}
}
\description{
this function checks to see whether any cohorts have been observed from birth until \code{OPENAGE} and returns a list with logical elements for males, females, etc. The object created by this function is used in the cohort lifetable functions, \code{ltcoh_AxN()} and \code{ltcohBoth_AxN()}, which use the argument \code{run.condition} to decide which list element (\code{"m"}, code{"f"}, \code{"both"}, \code{"either"}) to extract.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

