% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exposures_coh.R
\name{Exposures_coh}
\alias{Exposures_coh}
\title{\code{Exposures_coh} a function to prepare either v5 or v6 cohort exposures.}
\usage{
Exposures_coh(WORKING = getwd(), pop = NULL, dl = NULL, du = NULL,
  cohComp = NULL, sex, OPENAGE = 110, save.bin = TRUE, MPVERSION = 6,
  XXX = NULL, LDBPATH = NULL, IDBPATH = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD country abbreviation. Default \code{getwd()}.}

\item{pop}{optional. An 'age by year' matrix of Jan 1st population counts. Taken from \code{cohComp} or derived if necessary.}

\item{dl}{optional. An 'age by year' matrix of year t lower triangle death counts. Taken from \code{cohComp} or derived if necessary.}

\item{du}{optional. An 'age by year' matrix of year t+1 upper triangle death counts, territorially adjusted back to the year t standard. Taken from \code{cohComp} or derived if necessary.}

\item{cohComp}{optional. Output from \code{getCohortComponents()}.}

\item{sex}{\code{"m"} or \code{"f"}.}

\item{OPENAGE}{the desired open age. Default value is 110}

\item{save.bin}{logical. Default \code{TRUE}. Should the output be saved as e.g. \code{Rbin/cohExposuresRaw.Rdata} as well.}

\item{MPVERSION}{5 or 6. Version 5 exposures assume uniformity, v6 allows for non-uniformity across birthdays in the death distribution by taking information from monthly birth distributions. Differences are negligible.}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is extracted from \code{WORKING} as the last path part.}

\item{LDBPATH}{in case the LexisDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "LexisDB")}}

\item{IDBPATH}{in case the InputDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "InputDB")}}
}
\value{
an 'age by year' matrix of exposures. if in test mode, a list with matrices of both deaths and exposures is returned.
}
\description{
This function is meant to be internal. It is called by \code{cExposures_and_cMx_AxN()}, \code{ltcoh_AxN()} and \code{ltcohBoth_AxN()}. If necessary, it calls \code{getCohortComponents()}. Exposures from the appropriate MP version are returned in an 'age by year' matrix already cut down to \code{OPENAGE}, but with all cohorts present that are present in the uncut \code{pop}- i.e. cohort selection for lifetables or exposure writing needs to happen downstream.
}
\details{
This function can take data arguments in-memory, from the \code{/Rbin/} folder, or else derive them by calling \code{getCohortComponents()}. This function is separate because using monthly birth data for v6 exposures is a fair amount of code. Since a few functions need these exposures it's best to modularize and call where needed. This function uses a package dataset \code{monthDurations.rda} in order to save having to recompute it on each run- this is a matrix with the number of days in each month every year across several centuries, reaching way beyond the potential of the HMD to as to avoid future bugs. It gets cut to size as needed.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

