% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortComponents.R
\name{getCohortComponents}
\alias{getCohortComponents}
\title{\code{getCohortComponents} a function to read in and prepare basic data components for cohort calculations}
\usage{
getCohortComponents(WORKING = getwd(), sex, save.bin = TRUE, XXX = NULL,
  LDBPATH = NULL, IDBPATH = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD country abbreviation. Default \code{getwd()}.}

\item{sex}{either \code{"m"} or \code{"f"}}

\item{save.bin}{logical. Default \code{TRUE}. Should the output be saved as e.g. \code{Rbin/f_cohortComponents.Rdata.Rdata} / \code{Rbin/m_cohortComponents.Rdata.Rdata} as well? Appropriate name is derived automatically.}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is extracted from \code{WORKING} as the last path part.}

\item{LDBPATH}{in case the LexisDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "LexisDB")}}

\item{IDBPATH}{in case the InputDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "InputDB")}}
}
\value{
a long \code{data.frame} with 5 columns: \code{"Year"},\code{"Age"}, \code{"dl"}, \code{"du"} and \code{"pop"}.
}
\description{
This function modularizes the reading in, staggering, and post-territorial adjusting of the 3 basic (v5) data components of cohort calculations: pop, dl and du, and sticks them in a long-format \code{data.frame} amenable to reshaping in downstream functions.
}
\details{
This function is called by \code{Exposures_coh()}, \code{ltcoh_Axn()}, \code{ltcohBoth_AxN()} and \code{cExposures_and_cMx_AxN()} when needed, though these functions can also take the output of \code{getCohortComponents()} as input for in-memory data object passing (faster). The function \code{RunHMDCountry()} uses it this way. Another alternative is to save off the output of this function and read it in using \code{load()}, which is another built-in feature of the above functions. This function calls \code{cohTadj()} when pertinent.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

