% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHMDCountry.R
\docType{data}
\name{monthDurations}
\alias{monthDurations}
\title{A lookup table for the number of days in each month}
\format{A \code{matrix} with 12 rows and 401 columns}
\description{
An integer \code{matrix} containing the number of days in each month from 1700 until 2100.
Months (1-12) are in rows and years (1700-2100) are in columns. This \code{matrix} is used in
MPv6 exposure calculations only.
}
\keyword{datasets}

