% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perTadj.R
\name{perTadj}
\alias{perTadj}
\title{\code{perTadj} a function for doing lifetable territorial adjustments for period data}
\usage{
perTadj(LDBobj, tadj, sex)
}
\arguments{
\item{LDBobj}{Lexis database object, with proper column names, to be passed in by \code{getPeriodComponents()}.}

\item{tadj}{a territorial adjustment object, as read straight from the Input Database.}

\item{sex}{\code{"m"} or \code{"f"}}
}
\value{
returns LDBobj back again, having 'un'-adjusted deaths and pop counts for pertinent years.
}
\description{
period territorial adjustments 'un'adjust population counts bring January 1 population estimates from year t + 1 to December 31  of year t, since both populations are needed to calculate year t exposures. This step should in the future be moved back to the LDB programs, which should return a more comprehensive Lexis Database object. This function is called only by \code{getPeriodComponents()}.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

