% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population_A.R
\name{Population_A}
\alias{Population_A}
\title{\code{Population_A} a function to prepare data for the \code{Population} and \code{Population5} data objects}
\usage{
Population_A(WORKING = getwd(), abridged = FALSE, OPENAGE = 110,
  save.bin = TRUE, XXX = NULL, LDBPATH = NULL, IDBPATH = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD country abbreviation. Default \code{getwd()}.}

\item{abridged}{logical. Default \code{FALSE}.}

\item{OPENAGE}{the desired open age. Default value is 110.}

\item{save.bin}{logical. Default \code{TRUE}. Should the output be saved as e.g. \code{Rbin/Population.Rdata} as well? Appropriate name is derived systematically.}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is extracted from \code{WORKING} as the last path part.}

\item{LDBPATH}{in case the LexisDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "LexisDB")}}

\item{IDBPATH}{in case the InputDB is not in \code{WORKING} (local testing), the full path to the LexisDB folder. If left as \code{NULL} it is assumed to be \code{file.path(WORKING, "InputDB")}}
}
\value{
a \code{data.frame} in long (staced) format with \code{"Year"}, \code{"Age"}, \code{"Female"}, \code{"Male"} and \code{"Total"} columns. Values are unrounded.
}
\description{
This function is top level. It prepares population counts for males, females and Total, in the typical HMD output format (long, stacked). One peculiarity is that tadj years are stacked, with '-' and '+' appended to the year names. That makes this function slightly harder to integrate into the modularity of the other functions, a job which is not yet complete. For instance, it doesn't use \code{getPeriodExposures()}, but ideally would do so.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}
