% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLDB.R
\name{readLDB}
\alias{readLDB}
\title{function to read Lexis database files into R}
\usage{
readLDB(ldb.path)
}
\arguments{
\item{ldb.path}{path to the standard Lexis database file, including file extension \code{.txt}. File name should be of the form \code{"mESP.txt"} or \code{"fESP"}, where ESP could be any country code.}
}
\value{
data.frame LexisDB object, with columns in a useable form and a column for Sex added
}
\description{
This function reads in a given LexisDB file, imputing NAs properly (but not removing them), converting the Lexis column to character, and appending a Sex column, inferred from the first letter of the file name, per HMD conventions.
}
