% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_csv.R
\name{Write_csv}
\alias{Write_csv}
\title{\code{Write_csv} a function to write formatted .csv files containing
  1x1 lifetable data, plus occurances and exposures.}
\usage{
Write_csv(WORKING = getwd(), STATSFOLDER = "RSTATS", MPVERSION,
  OldStyle = TRUE, XXX = NULL, CountryLong = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD 
country abbreviation. Default \code{getwd()}.}

\item{STATSFOLDER}{the folder name where output is to be written to (not a 
full path). Default \code{"RSTATS"}.}

\item{MPVERSION}{5 or 6. Default 5. Here this only affects file headers.}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is
extracted from \code{WORKING} as the last path part.}

\item{CountryLong}{the HMD country full name.}

\item{OldStyle.}{Logical.  When true, output is .csv files analogous to
those produced under Matlab V5 code (no headers), that are part of the
input for SAS diagnostics.  When FALSE (default, not yet implemented) the output will be ....}
}
\value{
function called for its side effect of creating the lifetable txt
  output files, e.g. \code{mltper_1x1.txt} and other time/sex configurations.
  No value returned.
}
\description{
This function requires that all pertinent \code{ltper_AxN.Rdata}
  and \code{ltcoh_AxN()} data objects be present in the folder 
  \code{WORKING/Rbin/} for males, females and both-sex tables. Objects are 
  selected by grepping, which works for the present HMD file naming scheme. 
  \code{.Rdata} files are read in, rounded, formatted and written back out 
  into \code{WORKING/RSTATS/} by default, although this can be changed with 
  the \code{STATSFOLDER} argument.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}
