% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_Exposures_lexis.R
\name{Write_Exposures_lexis}
\alias{Write_Exposures_lexis}
\title{\code{Write_Exposures_lexis} a function to prepare and save the file \code{exposures_lexis.txt}.}
\usage{
Write_Exposures_lexis(WORKING = getwd(), STATSFOLDER = "RSTATS",
  BINFOLDER = "Rbin", BINPATH = NULL, OPENAGE = 110, MPVERSION,
  XXX = NULL, CountryLong = NULL)
}
\arguments{
\item{WORKING}{path to working directory, which typically ends with the HMD country abbreviation. Default \code{getwd()}.}

\item{BINFOLDER}{the folder name where binary output was written  (not a full path). Default \code{"Rbin"}.}

\item{OPENAGE}{the desired open age. Default value is 110.}

\item{MPVERSION}{5 or 6. Default 5. Here this only affects file headers.}

\item{XXX}{the HMD country abbreviation. If left \code{NULL}, this is extracted from \code{WORKING} as the last path part.}

\item{CountryLong}{the HMD country full name.}
}
\value{
function called for its side effect of creating the files \code{exposures_lexis.txt}. No value returned.
}
\description{
This function is essentially a stand-alone that prepares \code{exposures_lexis.txt} straight from binary files.
}
\author{
Carl Boe \email{boe@demog.berkeley.edu}
}
