% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearAgg.R
\name{YearAgg}
\alias{YearAgg}
\title{\code{YearAgg} Aggregate an 'age by year' matrix into N-year groups}
\usage{
YearAgg(mat, years = as.integer(colnames(mat)), N = 1)
}
\arguments{
\item{mat}{an 'age by year' numeric matrix, ideally with dimensions labeled.}

\item{years}{vector of years (columns) in \code{mat}. Defaults to the column names of \code{mat}}

\item{N}{the desired year-group width.}
}
\value{
an 'age by year' matrix with columns. Column labls will now specify hyphen-separated year ranges.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

